requirements.txt
streamlit==1.28.2
python-docx==1.1.0
tensorflow==2.15.0
tensorflow-hub==0.15.0
scikit-learn==1.3.2
PyPDF4==1.27.0
nltk==3.8.1
spacy==3.7.2